+++ HISTORY.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+2.0.3 (2016-12-20)
+------------------
+
+- Be able to pass in a site object to the render_content_core function for
+  layout aware items
+  [vangheem]
+
+
 2.0.2 (2016-12-14)
 ------------------